First mobile version

+++ Asteroids/Assets/Scripts/MonoBehaviours/Enemies/BaseEnemy.cs
@@ -7,7 +7,7 @@ public abstract class BaseEnemy : MonoBehaviour, IDamageable, IDamager
 
     public byte ScorePoints;
 
-    [SerializeField] protected ushort _speed;
+    [SerializeField] protected float _speed;
     [SerializeField] private AudioClip _dieSound;
 
     private void OnCollisionEnter2D(Collision2D collision)
